extract variables input
