+++ plugins/backstage-plugin-env0/src/components/env0-variables-input/index.ts
@@ -1,1 +1,2 @@
+export * from './env0-variables-scaffolder-input';
 export * from './env0-variables-input';
