+++ plugins/backstage-plugin-env0/src/hooks/use-variables-data-by-template.tsx
@@ -5,6 +5,7 @@ import { useAsyncRetry } from 'react-use';
 export const useVariablesDataByTemplate = (
   templateId?: string,
   projectId?: string,
+  environmentId?: string
 ) => {
   const apiClient = useApi<Env0Api>(env0ApiRef);
 
@@ -18,6 +19,7 @@ export const useVariablesDataByTemplate = (
     const template = await apiClient.getTemplateById(templateId);
 
     return await apiClient.listVariables({
+      environmentId,
       projectId,
       blueprintId: templateId,
       organizationId: template.organizationId,
